+++ x/stablecoin/types/params.go
@@ -7,6 +7,7 @@ import (
 	paramtypes "github.com/cosmos/cosmos-sdk/x/params/types"
 
 	"github.com/NibiruChain/nibiru/x/common"
+	epochstypes "github.com/NibiruChain/nibiru/x/epochs/types"
 )
 
 var _ paramtypes.ParamSet = (*Params)(nil)
@@ -57,13 +58,12 @@ func DefaultParams() Params {
 	feeRatio := sdk.MustNewDecFromStr("0.002")
 	efFeeRatio := sdk.MustNewDecFromStr("0.5")
 	bonusRateRecoll := sdk.MustNewDecFromStr("0.002")
-	distrEpochIdentifier := "15 min"
 	adjustmentStep := sdk.MustNewDecFromStr("0.0025")
 	priceLowerBound := sdk.MustNewDecFromStr("0.9999")
 	priceUpperBound := sdk.MustNewDecFromStr("1.0001")
 	isCollateralRatioValid := false // Will be valid once we start posting prices and updating the collateral
 
-	return NewParams(genesisCollRatio, feeRatio, efFeeRatio, bonusRateRecoll, distrEpochIdentifier,
+	return NewParams(genesisCollRatio, feeRatio, efFeeRatio, bonusRateRecoll, epochstypes.FifteenMinuteEpochID,
 		adjustmentStep,
 		priceLowerBound,
 		priceUpperBound, isCollateralRatioValid)